refactor(scripts): restructure prepare-release code for reuse
Extract common functionality from prepare-release script into separate
modules, allowing reuse in new pack-as-beta script. Improve organization
with dedicated files for specific functions such as scope changes,
dependency updates, and file traversal.

Issue: BTC-1933
